Introduces new function to "common" script for Steam Guard authentication
Multi-shot refinement of proposed solution that sketched out support for Steam Guard via a simple "shim" ahead of the invocation of Supervisor (as 'supervisord' will ultimately curtail "interactive" nature of the shell).

The function 'steamGuardAuth' is only fully traversed if the environment variable 'USE_STEAMGUARD' is "true". If "false", the routine returns.. In either scenario, control is ultimately yielded back to Supervisor - engaging the Supervisor program for "starbound-updater" and "starbound-bootstrap".

While in the function, the user has a configurable amount of attempts to log in with Steam Guard successfully ('STEAMGUARD_MAX_RETRIES' environment variable, default of "3"). If the user fails all attempts, the overall deployment attempt is halted (exit).

If a Steam Guard login attempt works, the remainder of the "machine" kicks in - all of the typical install/update/check routines.