Changes the 'restart' policy in the 'starbound' service to "on-failure:3"
Building on the presence of 'STEAMGUARD_TIMEOUT', the SteamCMD authentication routine will either pass (i.e. successful auth-n using Steam Credentials and Steam Guard code if applicable) or fail (i.e. invalid Steam Credentials and/or Steam Guard code). When the "fail" condition happens, the 'steamGuardAuth' function emits 'exit 2' - which ultimately yields "what happens next" (in terms of the container itself) to the container service's 'restart' policy - previously "unlress-stopped", which would effectively result in an "infinite" loop until of container relaunch and therefore "bootstrap" script execution and therefore SteamCMD auth-n and therefore (fail) and so on ...

Setting "on-failure:3" in the 'starbound' service in Docker Compose will effectively curtail the aforementioned loop as it is not desirable for the remainder of the "bootstrap" script to proceed until auth-n is successful. This, coupled with a reasonable 'STEAMGUARD_TIMEOUT' value will create the opportunity to successfully resolve Steam authentication (and Steam Guard code input, if applicable) - and stop the container if this routine fails too many times.