Introduces a hard exit (0) to stop the container when Steam Guard authentication times out or ultimately fails (as managed by SteamCMD)
While the container 'restart' policy is overall authoritative, the early "bootstrap" process invokes a function ('steamGuardAuth') that is effectively a hard pass/fail outcome.

Consider a ;restart' policy of "unless-stopped". Steam Guard auth-n that fails/quits/times out would result in the container restarting - and repeating the process ad infinitum.

Consider a 'restart' policy of "on-failure:3". Steam Guard auth-n that fails/quits/times out would result in the container restarting - and counting against the allocated "failures" for the remainder of the lifecycle of the container (even after potential success of the Steam Guard auth-n process).

When considering either scenario, the more risk-prone outcome prevails: effectively a "silent" failure that would require intervention when noticed ("unlress-stopped") or eventually ("on-failure:3").

Introducing this hard exit for the container process at this one specific and very-early stage more effectively enforces a pass/fail constraint - and is more likely to yield an immediate response from container orchestration/alerting/admin.