Adds timezone-override support to container environment
Included 'tzdata' package in Dockerfile package manifest. Added "TZ" environment key in Docker Compose with initial value of "Etc/UTC" (which is the default even when not defined).

Added a friendly default entry in "defaults" script for scenarios where the environment is somehow reset (not encountered yet, so this will be considered a failover)