Adds cron-validator installation to Dockerfile
Installed 'cron-validator' in preparation for more robust validation of cron schedules.

In terms of using 'break_system_packages', the container image is single-purpose - and it is not expected that anyone making use of this image in context of this repo will 'pip install ...' inside the running container.

In this very-specific scenario, it is acceptable to push the 'cron-validator' package into system-managed space since Dockerfile and "bootstrap" script are "static" once invoked - i.e. nothing else is expected to be installed beyond what is already outlined ahead of time.